+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- chore: improve error logging when capturing snapshots ([#146](https://github.com/PostHog/posthog-flutter/pull/146))
+
 ## 4.9.1
 
 - fix: blank screen when viewing session replay recordings ([#139](https://github.com/PostHog/posthog-flutter/pull/139))